+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `/api/v1/search`
   - `/api/v1/statuses/{id}/card` 
 - LDAP authenticator
+- Chats, they were half-baked. Just use PMs.
 
 ## 2022.07